Check your nlmixr2 installation for potential issues

Check your nlmixr2 installation for potential issues

Usage

nlmixr2CheckInstall()

Examples

nlmixr2CheckInstall()
#> ℹ Operating system: Linux 6.11.0-1018-azure #18~24.04.1-Ubuntu SMP Sat Jun 28 04:46:03 UTC 2025
#> Loading required namespace: devtools
#> ✔ The 'make' command was found:  /usr/bin/make
#> ℹ GNU Make 4.3
#> Built for x86_64-pc-linux-gnu
#> Copyright (C) 1988-2020 Free Software Foundation, Inc.
#> License GPLv3+: GNU GPL version 3 or later <http://gnu.org/licenses/gpl.html>
#> This is free software: you are free to change and redistribute it.
#> There is NO WARRANTY, to the extent permitted by law.
#> ✔ The package 'rxode2' is installed and seems to be up to date, version 4.0.0
#> ✔ The package 'rxode2ll' is installed and seems to be up to date, version 2.0.13
#> ✔ The package 'lotri' is installed and seems to be up to date, version 1.0.0.9000
#> ✔ The package 'nlmixr2' is installed and seems to be up to date, version 4.0.0
#> ✔ The package 'nlmixr2est' is installed and seems to be up to date, version 4.0.0
#> ✔ The package 'nlmixr2data' is installed and seems to be up to date, version 2.0.9
#> ✔ The package 'nlmixr2extra' is installed and seems to be up to date, version 3.0.3
#> ✔ The package 'nlmixr2plot' is installed and seems to be up to date, version 3.0.2
#> ✖ The package 'babelmixr2' is not installed (it is optional for all rxode2/nlmixr2 work)
#> ✔ The package 'ggPMX' is installed and seems to be up to date, version 1.3.0
#> ✖ The package 'monolix2rx' is not installed (it is optional for all rxode2/nlmixr2 work)
#> ✖ The package 'nlmixr2lib' is not installed (it is optional for all rxode2/nlmixr2 work)
#> ✖ The package 'nlmixr2rpt' is not installed (it is optional for all rxode2/nlmixr2 work)
#> ✖ The package 'nonmem2rx' is not installed (it is optional for all rxode2/nlmixr2 work)
#> ✖ The package 'shinyMixR' is not installed (it is optional for all rxode2/nlmixr2 work)
#> ✔ The package 'xpose.nlmixr2' is installed and seems to be up to date, version 0.4.1
#> ℹ To install missing packages, run the following command:
#> install.packages(c('babelmixr2', 'monolix2rx', 'nlmixr2lib', 'nlmixr2rpt', 'nonmem2rx', 'shinyMixR'))