Estimate the objective function values for a model while fixing defined parameter values

Estimate the objective function values for a model while fixing defined parameter values

Usage

profileFixedSingle(fitted, which)

Arguments

fitted

The fit model

which

A data.frame with column names of parameters to fix and values of the fitted value to fix (one row only).

Value

which with a column named OFV added with the objective function value of the fitted estimate fixing the parameters in the other columns

Functions

profileFixedSingle(): Estimate the objective function value for a model while fixing a single set of defined parameter values (for use in parameter profiling)

See also