VPC simulation

VPC simulation

Usage

vpcSim(
  object,
  ...,
  keep = NULL,
  n = 300,
  pred = FALSE,
  seed = 1009,
  nretry = 50,
  minN = 10,
  normRelated = TRUE
)

Arguments

object

This is the nlmixr2 fit object

...

Additional arguments passed to nlmixr2est::vpcSim().

keep

Column names to keep in the output simulated dataset

n

Number of simulations

pred

Should predictions be added to the simulation

seed

Seed to set for the VPC simulation

nretry

Number of times to retry the simulation if there is NA values in the simulation

minN

With retries, the minimum number of studies to restimulate (by default 10)

normRelated

should the VPC style simulation be for normal related variables only

Value

data frame of the VPC simulation

Author

Matthew L. Fidler

Examples

# \donttest{

if (rxode2::.linCmtSensB()) {

one.cmt <- function() {
 ini({
   ## You may label each parameter with a comment
   tka <- 0.45 # Log Ka
   tcl <- log(c(0, 2.7, 100)) # Log Cl
   ## This works with interactive models
   ## You may also label the preceding line with label("label text")
   tv <- 3.45; label("log V")
   ## the label("Label name") works with all models
   eta.ka ~ 0.6
   eta.cl ~ 0.3
   eta.v ~ 0.1
   add.sd <- 0.7
 })
 model({
   ka <- exp(tka + eta.ka)
   cl <- exp(tcl + eta.cl)
   v <- exp(tv + eta.v)
   linCmt() ~ add(add.sd)
 })
}

fit <- nlmixr(one.cmt, theo_sd, est="focei")

head(vpcSim(fit, pred=TRUE))

}
#>  
#>  
#>  
#>  
#> ℹ parameter labels from comments are typically ignored in non-interactive mode
#> ℹ Need to run with the source intact to parse comments
#> → Calculating residuals/tables
#> ✔ done
#> → compress origData in nlmixr2 object, save 5952
#> → compress parHistData in nlmixr2 object, save 4856
#>  
#>  
#>   sim.id id time    ipred        sim  tad nlmixrRowNums rxLambda rxYj rxLow
#> 1      1  1 0.00 0.000000 -0.2603687 0.00             2        1    2     0
#> 2      1  1 0.25 4.241461  3.8578359 0.25             3        1    2     0
#> 3      1  1 0.57 6.819229  7.1918839 0.57             4        1    2     0
#> 4      1  1 1.12 8.177318  7.2241716 1.12             5        1    2     0
#> 5      1  1 2.02 8.192648  8.2314881 2.02             6        1    2     0
#> 6      1  1 3.82 7.366554  7.6343568 3.82             7        1    2     0
#>   rxHi     pred
#> 1    1 0.000000
#> 2    1 3.270856
#> 3    1 5.839073
#> 4    1 7.868363
#> 5    1 8.501092
#> 6    1 7.615773

# }