Stepwise Covariate Model-selection (SCM) method

Stepwise Covariate Model-selection (SCM) method

Usage

covarSearchAuto(
  fit,
  varsVec,
  covarsVec,
  pVal = list(fwd = 0.05, bck = 0.01),
  covInformation = NULL,
  catCovariates = NULL,
  searchType = c("scm", "forward", "backward"),
  restart = FALSE
)

Arguments

fit

an nlmixr2 'fit' object

varsVec

a list of candidate variables to which the covariates could be added

covarsVec

a list of candidate covariates that need to be tested

pVal

a named list with names 'fwd' and 'bck' for specifying the p-values for the forward and backward searches, respectively

covInformation

a list containing additional information on the variables-covariates pairs that should be passed on to addCovMultiple function

catCovariates

a list of covariates that should be treated as categorical

searchType

one of 'scm', 'forward' and 'backward' to specify the covariate search method; default is 'scm'

restart

a boolean that controls if the search should be restarted; default is FALSE

Value

A list summarizing the covariate selection steps and output; This list has the "summaryTable" for the overall summary of the covariate selection as well as "resFwd" for the forward selection method and "resBck" for the backward selection method.

Author

Vipul Mann, Matthew Fidler

Examples

# \donttest{

one.cmt <- function() {
  ini({
    ## You may label each parameter with a comment
    tka <- 0.45 # Log Ka
    tcl <- log(c(0, 2.7, 100)) # Log Cl
    ## This works with interactive models
    ## You may also label the preceding line with label("label text")
    tv <- 3.45; label("log V")
    ## the label("Label name") works with all models
    eta.ka ~ 0.6
    eta.cl ~ 0.3
    eta.v ~ 0.1
    add.sd <- 0.7
  })
  model({
    ka <- exp(tka + eta.ka)
    cl <- exp(tcl + eta.cl)
    v <- exp(tv + eta.v)
    linCmt() ~ add(add.sd)
  })
}

fit <- nlmixr2(one.cmt, nlmixr2data::theo_sd,"focei")
#>  
#>  
#>  
#> ℹ parameter labels from comments will be replaced by 'label()'
#> → Calculating residuals/tables
#> ✔ done
#> → compress origData in nlmixr2 object, save 5952
#> → compress parHist in nlmixr2 object, save 2072
#> Warning: gradient problems with initial estimate and covariance; see $scaleInfo
#> Warning: last objective function was not at minimum, possible problems in optimization
#> Warning: ETAs were reset to zero during optimization; (Can control by foceiControl(resetEtaP=.))
#> Warning: initial ETAs were nudged; (can control by foceiControl(etaNudge=., etaNudge2=))
rxode2::.rxWithWd(tempdir(), {# with temporary directory

auto1 <- covarSearchAuto(fit, varsVec = c("ka", "cl"),
    covarsVec = c("WT", "SEX"), catCovariates = c("SEX"))

})
#> Warning: some covariates are not in the dataset: 'WT', ''
#> 
#> ── starting forward search... ──────────────────────────────────────────────────
#> ✔ added centered_WT*cov_WT_ka to ka's equation in the model
#> ✔ updated value: ka<-exp(tka+centered_WT*cov_WT_ka+eta.ka)
#> ℹ add population parameter `centered_WT` and set estimate to 1
#> ℹ add population parameter `cov_WT_ka` and set estimate to 1
#>  
#>  
#> ℹ use `control` from pipeline
#> → loading into symengine environment...
#> → pruning branches (`if`/`else`) of full model...
#> ✔ done
#> → calculate jacobian
#> → calculate ∂(f)/∂(η)
#> → calculate ∂(R²)/∂(η)
#> → finding duplicate expressions in inner model...
#> → optimizing duplicate expressions in inner model...
#> → finding duplicate expressions in EBE model...
#> → optimizing duplicate expressions in EBE model...
#> → compiling inner model...
#>  
#> ✔ done
#> → finding duplicate expressions in FD model...
#> → compiling EBE model...
#>  
#> ✔ done
#> → compiling events FD model...
#>  
#> ✔ done
#> → Calculating residuals/tables
#> ✔ done
#> → compress origData in nlmixr2 object, save 7072
#> → compress parHist in nlmixr2 object, save 2800
#> ✔ added centered_WT*cov_WT_cl to cl's equation in the model
#> ✔ updated value: cl<-exp(tcl+centered_WT*cov_WT_cl+eta.cl)
#> ℹ add population parameter `centered_WT` and set estimate to 1
#> ℹ add population parameter `cov_WT_cl` and set estimate to 1
#>  
#>  
#> ℹ use `control` from pipeline
#> → loading into symengine environment...
#> → pruning branches (`if`/`else`) of full model...
#> ✔ done
#> → calculate jacobian
#> → calculate ∂(f)/∂(η)
#> → calculate ∂(R²)/∂(η)
#> → finding duplicate expressions in inner model...
#> → optimizing duplicate expressions in inner model...
#> → finding duplicate expressions in EBE model...
#> → optimizing duplicate expressions in EBE model...
#> → compiling inner model...
#>  
#> ✔ done
#> → finding duplicate expressions in FD model...
#> → compiling EBE model...
#>  
#> ✔ done
#> → compiling events FD model...
#>  
#> ✔ done
#> → Calculating residuals/tables
#> ✔ done
#> → compress origData in nlmixr2 object, save 7072
#> → compress parHist in nlmixr2 object, save 2400
#> 
#> ── objf value did not improve, exiting the search ... ──────────────────────────
#> 
#> ── forward search complete ──
#> 
#> ── starting backward search... ─────────────────────────────────────────────────
#> ! no covariates added in the forward search, skipping backward search

## Note that this didn't include sex, add it to dataset and restart model

d <- nlmixr2data::theo_sd
d$SEX <-0
d$SEX[d$ID<=6] <-1

fit <- nlmixr2(one.cmt, d, "focei")
#>  
#>  
#>  
#> ℹ parameter labels from comments will be replaced by 'label()'
#> ℹ use `control` from pipeline
#> → Calculating residuals/tables
#> ✔ done
#> → compress origData in nlmixr2 object, save 7184
#> → compress parHist in nlmixr2 object, save 2072
#> Warning: gradient problems with initial estimate and covariance; see $scaleInfo
#> Warning: last objective function was not at minimum, possible problems in optimization
#> Warning: ETAs were reset to zero during optimization; (Can control by foceiControl(resetEtaP=.))
#> Warning: initial ETAs were nudged; (can control by foceiControl(etaNudge=., etaNudge2=))

# This would restart if for some reason the search crashed:

rxode2::.rxWithWd(tempdir(), {# with temporary directory

auto2 <- covarSearchAuto(fit, varsVec = c("ka", "cl"), covarsVec = c("WT", "SEX"),
                catCovariates = c("SEX"), restart = TRUE)

auto3 <- covarSearchAuto(fit, varsVec = c("ka", "cl"), covarsVec = c("WT", "SEX"),
                catCovariates = c("SEX"), restart = TRUE,
                searchType = "forward")
})
#> 
#> ── starting forward search... ──────────────────────────────────────────────────
#> ✔ added centered_WT*cov_WT_ka to ka's equation in the model
#> ✔ updated value: ka<-exp(tka+centered_WT*cov_WT_ka+eta.ka)
#> ℹ add population parameter `centered_WT` and set estimate to 1
#> ℹ add population parameter `cov_WT_ka` and set estimate to 1
#>  
#>  
#> ℹ use `control` from pipeline
#> → Calculating residuals/tables
#> ✔ done
#> → compress origData in nlmixr2 object, save 8432
#> → compress parHist in nlmixr2 object, save 2800
#> ✔ added centered_WT*cov_WT_cl to cl's equation in the model
#> ✔ updated value: cl<-exp(tcl+centered_WT*cov_WT_cl+eta.cl)
#> ℹ add population parameter `centered_WT` and set estimate to 1
#> ℹ add population parameter `cov_WT_cl` and set estimate to 1
#>  
#>  
#> ℹ use `control` from pipeline
#> → Calculating residuals/tables
#> ✔ done
#> → compress origData in nlmixr2 object, save 8432
#> → compress parHist in nlmixr2 object, save 2400
#> ✔ added categorical_SEX_0*cov_SEX_0_ka to ka's equation in the model
#> ✔ updated value: ka<-exp(tka+categorical_SEX_0*cov_SEX_0_ka+eta.ka)
#> ℹ add covariate `categorical_SEX_0`
#> ℹ add population parameter `cov_SEX_0_ka` and set estimate to 1
#>  
#>  
#> ℹ use `control` from pipeline
#> → loading into symengine environment...
#> → pruning branches (`if`/`else`) of full model...
#> ✔ done
#> → calculate jacobian
#> → calculate ∂(f)/∂(η)
#> → calculate ∂(R²)/∂(η)
#> → finding duplicate expressions in inner model...
#> → optimizing duplicate expressions in inner model...
#> → finding duplicate expressions in EBE model...
#> → optimizing duplicate expressions in EBE model...
#> → compiling inner model...
#>  
#> ✔ done
#> → finding duplicate expressions in FD model...
#> → compiling EBE model...
#>  
#> ✔ done
#> → compiling events FD model...
#>  
#> ✔ done
#> → Calculating residuals/tables
#> ✔ done
#> → compress origData in nlmixr2 object, save 7984
#> → compress parHist in nlmixr2 object, save 2816
#> ✔ added categorical_SEX_0*cov_SEX_0_cl to cl's equation in the model
#> ✔ updated value: cl<-exp(tcl+categorical_SEX_0*cov_SEX_0_cl+eta.cl)
#> ℹ add covariate `categorical_SEX_0`
#> ℹ add population parameter `cov_SEX_0_cl` and set estimate to 1
#>  
#>  
#> ℹ use `control` from pipeline
#> → loading into symengine environment...
#> → pruning branches (`if`/`else`) of full model...
#> ✔ done
#> → calculate jacobian
#> → calculate ∂(f)/∂(η)
#> → calculate ∂(R²)/∂(η)
#> → finding duplicate expressions in inner model...
#> → optimizing duplicate expressions in inner model...
#> → finding duplicate expressions in EBE model...
#> → optimizing duplicate expressions in EBE model...
#> → compiling inner model...
#>  
#> ✔ done
#> → finding duplicate expressions in FD model...
#> → compiling EBE model...
#>  
#> ✔ done
#> → compiling events FD model...
#>  
#> ✔ done
#> → Calculating residuals/tables
#> ✔ done
#> → compress origData in nlmixr2 object, save 7984
#> → compress parHist in nlmixr2 object, save 2648
#> 
#> ── objf value did not improve, exiting the search ... ──────────────────────────
#> 
#> ── forward search complete ──
#> 
#> ── starting backward search... ─────────────────────────────────────────────────
#> ! no covariates added in the forward search, skipping backward search
#> 
#> ── starting forward search... ──────────────────────────────────────────────────
#> ✔ added centered_WT*cov_WT_ka to ka's equation in the model
#> ✔ updated value: ka<-exp(tka+centered_WT*cov_WT_ka+eta.ka)
#> ℹ add population parameter `centered_WT` and set estimate to 1
#> ℹ add population parameter `cov_WT_ka` and set estimate to 1
#>  
#>  
#> ℹ use `control` from pipeline
#> → Calculating residuals/tables
#> ✔ done
#> → compress origData in nlmixr2 object, save 8432
#> → compress parHist in nlmixr2 object, save 2800
#> ✔ added centered_WT*cov_WT_cl to cl's equation in the model
#> ✔ updated value: cl<-exp(tcl+centered_WT*cov_WT_cl+eta.cl)
#> ℹ add population parameter `centered_WT` and set estimate to 1
#> ℹ add population parameter `cov_WT_cl` and set estimate to 1
#>  
#>  
#> ℹ use `control` from pipeline
#> → Calculating residuals/tables
#> ✔ done
#> → compress origData in nlmixr2 object, save 8432
#> → compress parHist in nlmixr2 object, save 2400
#> ✔ added categorical_SEX_0*cov_SEX_0_ka to ka's equation in the model
#> ✔ updated value: ka<-exp(tka+categorical_SEX_0*cov_SEX_0_ka+eta.ka)
#> ℹ add covariate `categorical_SEX_0`
#> ℹ add population parameter `cov_SEX_0_ka` and set estimate to 1
#>  
#>  
#> ℹ use `control` from pipeline
#> → Calculating residuals/tables
#> ✔ done
#> → compress origData in nlmixr2 object, save 7984
#> → compress parHist in nlmixr2 object, save 2816
#> ✔ added categorical_SEX_0*cov_SEX_0_cl to cl's equation in the model
#> ✔ updated value: cl<-exp(tcl+categorical_SEX_0*cov_SEX_0_cl+eta.cl)
#> ℹ add covariate `categorical_SEX_0`
#> ℹ add population parameter `cov_SEX_0_cl` and set estimate to 1
#>  
#>  
#> ℹ use `control` from pipeline
#> → Calculating residuals/tables
#> ✔ done
#> → compress origData in nlmixr2 object, save 7984
#> → compress parHist in nlmixr2 object, save 2648
#> 
#> ── objf value did not improve, exiting the search ... ──────────────────────────
#> 
#> ── forward search complete ──
#> 

# }