VPC simulation

VPC simulation

Usage

vpcSim(
  object,
  ...,
  keep = NULL,
  n = 300,
  pred = FALSE,
  seed = 1009,
  nretry = 50
)

Arguments

object

This is the nlmixr2 fit object

...

Additional arguments passed to nlmixr2est::vpcSim().

keep

Keep character vector

n

Number of simulations

pred

Should predictions be added to the simulation

seed

Seed to set for the VPC simulation

nretry

Number of times to retry the simulation if there is NA values in the simulation

Value

data frame of the VPC simulation

Author

Matthew L. Fidler

Examples

# \donttest{
one.cmt <- function() {
 ini({
   ## You may label each parameter with a comment
   tka <- 0.45 # Log Ka
   tcl <- log(c(0, 2.7, 100)) # Log Cl
   ## This works with interactive models
   ## You may also label the preceding line with label("label text")
   tv <- 3.45; label("log V")
   ## the label("Label name") works with all models
   eta.ka ~ 0.6
   eta.cl ~ 0.3
   eta.v ~ 0.1
   add.sd <- 0.7
 })
 model({
   ka <- exp(tka + eta.ka)
   cl <- exp(tcl + eta.cl)
   v <- exp(tv + eta.v)
   linCmt() ~ add(add.sd)
 })
}

fit <- nlmixr(one.cmt, theo_sd, est="focei")
#>  
#>  
#>  
#> ℹ parameter labels from comments will be replaced by 'label()'
#> → Calculating residuals/tables
#> ✔ done
#> → compress origData in nlmixr2 object, save 5952
#> → compress parHist in nlmixr2 object, save 2072
#> Warning: gradient problems with initial estimate and covariance; see $scaleInfo
#> Warning: last objective function was not at minimum, possible problems in optimization
#> Warning: ETAs were reset to zero during optimization; (Can control by foceiControl(resetEtaP=.))
#> Warning: initial ETAs were nudged; (can control by foceiControl(etaNudge=., etaNudge2=))

head(vpcSim(fit, pred=TRUE))
#>  
#>   sim.id id time    ipred        sim nlmixrRowNums rxLambda rxYj rxLow rxHi
#> 1      1  1 0.00 0.000000 -0.6283886             2        1    2     0    1
#> 2      1  1 0.25 3.019619  2.6545739             3        1    2     0    1
#> 3      1  1 0.57 5.628867  6.6265894             4        1    2     0    1
#> 4      1  1 1.12 8.039087  8.7118336             5        1    2     0    1
#> 5      1  1 2.02 9.211883  8.9025113             6        1    2     0    1
#> 6      1  1 3.82 8.607713  7.9094706             7        1    2     0    1
#>       pred
#> 1 0.000000
#> 2 3.262732
#> 3 5.830087
#> 4 7.865576
#> 5 8.506533
#> 6 7.622375

# }